modified docstring
